Refactor modules (#74)
* Use namedtuple Transition for ease of code change and usage

* Move custom problems and agents under contrib module

* Fix contrib torchrl imports

* Add tests for experiments, fix updated runner

* Fix lint errors

+++ torchrl/contrib/agents/dqn_agent.py
@@ -1,11 +1,11 @@
 from copy import deepcopy
 import torch
 from torch.optim import Adam
+from torchrl.agents import BaseAgent
+from torchrl.policies import epsilon_greedy
+from torchrl.utils import ExpDecaySchedule
 
-from .base_agent import BaseAgent
-from ..policies import epsilon_greedy
 from ..models import QNet
-from ..utils import ExpDecaySchedule
 
 
 class BaseDQNAgent(BaseAgent):
@@ -59,7 +59,7 @@ class BaseDQNAgent(BaseAgent):
     return actions
 
   def compute_q_values(self, obs, action, reward, next_obs, done):
-    current_q_values = self.q_net(obs).gather(1, action)
+    current_q_values = self.q_net(obs).gather(1, action.long())
 
     with torch.no_grad():
       if self.double_dqn:
